show Content Block update events on in progress Hosts
Previously an event would only show in the
Edition's "history and notes" if the event took
place after the Edition had been published (and if
it was not archived/superseded).

We now want to display event if it took place at
any time after an Edition's creation, IF it is not
archived.