Migrate fact-check-email API to GmailV1
Google removed traditional auth methods for accessing inboxes, migrating to OAuth2.
The result of this is that we have to migrate our method of accessing the inboxes from
the original username/password system to a service account calling the GMail API.

Notable changes
- We no longer need MailFetcherConfig as the Mail gem is no longer used in the script
workflow. This config was only ever used in our automated script and is replaced
by the Gmail auth setup in FactCheckEmailHandler

- We no longer use email and password, instead using a JSON API secrets key that
is stored in AWS secrets and called via ENV.

Functionally the resulting logic is the same as previously.