Pluralize the text that displays the number of documents in the publications index page.
Make the number of documents displayed dynamic which shows singular or plural depending on the number of results.

Format the display of number of documents, so that it displays quantity over 1,000 with a comma separator.

Make number of documents over 1,000 display with a comma separating the thousand values using number_with_delimiter