Fix flakey test in PrimaryListingPresenterTest
The test only seems to care that the same items are returned,
irregardless of order (since the sorting for both arrays is done in the
test). Since all three editions that are created will have the same,
default, title, sorting by title is non-deterministic—use the
`assert_same_elements` matcher instead of sorting and using the
`assert_equal` one.