Add/move assertions to improve test reliablility
With more recent versions of Chromium, control is sometimes returning to
tests before the chromium action (such as visiting a page) has
completed. Therefore it is necessary to check that the action has
completed (for example, by asserting on some page content that appears
once the action is done) before continuing in the test.