Move tagging_mainstream_browse_page
Move the code for rendering the page to select a GOV.UK breadcrumb to
set from the `EditionsController` to the `TaggingController`.

This is part of moving the tagging-related code out of the
`EditionsController` to reduce the complexity and size of that
controller.

Also adds `.html` into the template's file name to be consistent with
the other template file names.