Upgrade to Rails 7.2
Of note is a change to validation on autosaving. Previously, validation
would occur on autosave whether or not a value had changed. However as
of 7.2.3, this behaviour was changed and no longer validates unchanged
values.

Practically this means that calling `publish!` alone on a saved document
no longer validates that document. There is no case where we do not
eventually validate all documents in save, so this is a safe behavioural
change, but one we need to be aware of.

Changenotes for 8.0.2 indicates this is a bug, so the changed test
may need to be reverted on further upgrade.