Update rails to 8.0.3
Previously we had locked to 8.0.2.1 due to an incompatibility with
GovSpeak and 8.0.3. This was caused due to GovSpeak being locked
to an older version due to an issue causing Publisher to fail to
load on Docker.

After the 8.0.2.1 update, I tested Publisher on docker with this
version lock removed, and it now works fine locally. Ergo we can bring
the tool up to date.