Extract common code to update_tags method
Since both `remove_breadcrumb` and `update_breadcrumb` were shared a lot
of the same code, extract that code to a private method in the
controller.

Due to the need to pass a block, it does make the code a bit more
complicated/tricky to read, but I think that is worth it to avoid too
much duplication (which will only get worse as methods for the other
tagging values are created).