Separate related content update code
Update the code related to updating the related content so that the
view does not need to have lots of hidden fields containing the other
tagging values (which is both a maintenance burden, and a potential
security issue).

The only hidden field now is the "previous_version", which is needed in
the case of multiple people editing the tags at the same time and
prevents one user overwriting another's changes.

When `Tagging::TaggingUpdateForm#publish!` is called, it will now raise
an `ActiveModel::ValidationError` if any of the `ordered_related_items`
don't have content IDs. This avoids the need to manually call the
`valid?` method and allows the exception to be caught in the one update
method that sets values that might trigger a validation failure
(`TaggingController#update_related_content`), which can then render the
correct page for the validation error to be displayed. Previously, there
was an assumption in `update_tagging` that the validation only failed
for related content changes, and so that is the page that should be
rendered if validation fails.

The line raising the validation error disables a rubocop cop because we
don't need to specify a validation error message.