Separate organisation update code
Update the code related to updating organisations so that
the view does not need to have lots of hidden fields containing the
other tagging values (which is both a maintenance burden, and a
potential security issue).

The only hidden field now is the "previous_version", which is needed in
the case of multiple people editing the tags at the same time and
prevents one user overwriting another's changes.