Remove unused code
Now that the `update_tagging` action of the `TaggingController` has been
replaced with separate methods for each type of tagging update, the
method can be removed.

One associated test has been removed, and another updated to work as
expected (it wasn't correctly mocking a method call and so was not
actually testing what it was supposed to be testing).